streamlit/s_requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
sentence-transformers

